+++ packages/create-theme/public/template/docs/.vitepress/theme/index.ts
@@ -1,4 +1,6 @@
 import BlogTheme from '@sugarat/theme'
+// 默认颜色仍然使用绿色
+import './green-theme.var.css'
 // 自定义样式重载
 // import './style.scss'