+++ packages/theme/docs/.vitepress/theme/index.ts
@@ -1,6 +1,6 @@
 import BlogTheme from '@sugarat/theme'
 import { enhanceAppWithTabs } from 'vitepress-plugin-tabs/client'
-
+import './green-theme.var.css'
 export default {
   ...BlogTheme,
   enhanceApp(ctx: any) {