+++ packages/theme/src/components/BlogRecommendArticle.vue
@@ -50,13 +50,13 @@
 import { ref, computed } from 'vue'
 import { useRoute, withBase } from 'vitepress'
 import { ElButton, ElLink } from 'element-plus'
-import { formatShowDate } from '../utils/index'
+import { formatShowDate } from '../utils/client'
 import { useArticles, useBlogConfig } from '../composables/config/blog'
 
 const { recommend: _recommend } = useBlogConfig()
 
 const sidebarStyle = computed(() =>
-  _recommend && _recommend?.style ? _recommend.style : 'card'
+  _recommend && _recommend?.style ? _recommend.style : 'sidebar'
 )
 
 const recommendPadding = computed(() =>